// src/api/openapi/models/AuditLog.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AuditLog = {
    id?: string;
    action?: string;
    createdat?: string;
    userid?: string;
    section?: string;
    avatar?: string;
    firstname?: string;
    lastname?: string;
    username?: string;
    role?: string;
    email?: string;
};